Add attack mechanic and rotation correction
Add a new method in InputReader to read mouse click. This method is
called in controller to check if the player wants to attack and further
conditionals is used to be sure the attack animation will be called only
once and wont overlap itself.

Also add a method in PlayerController to correct the Z rotation of the
player to zero by lerping its rotation when there arent new inputs.